Factor out http status from Runtime into its own file.
Rename the tailer httpstatus file for consistency.
